Lots of changes:
Removed elements from the quad tree, leaving only quad tree nodes.  It is enough for the purposes of terrain visualization, and simplifies the code.

Started implementing block management, blocks are subdivided or added when the camera moves.

Added possibility to listen to frame occurence, and added a fps counter.

Implemented a texture provider that renders the geotools map to a texture.

git-svn-id: https://svn.osgeo.org/geotools/geotools/trunk@26475 e5c1c795-43da-0310-a71f-fac65c449510