In referencing module, changed the DirectPosition2D.hashCode() and equals(Object) implementation in order to comply with DirectPosition contract, which fix an AssertionError. It has the unfortunate side-effect of violating the Point2D.hashCode() contract. Unfortunatly it is not possible to comply to both in same time. I added a note in javadoc trying to make that clear.
In the coverage module, applied the ImageWorker changes mentioned on the mailing list. Also provided a first draft of an ImageWorkerTest.

Note that the build still fail in the shapefile-renderer module for a reason that I believe unrelated to those changes. The same changes have already been applied to the 2.4 branch, which build fine.




git-svn-id: https://svn.osgeo.org/geotools/geotools/trunk@28779 e5c1c795-43da-0310-a71f-fac65c449510